refactor: 카드 패키지 생성
