+++ src/test/java/blackjack/domain/card/CardTest.java
@@ -1,4 +1,4 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
